# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PC Builder API",
    "contact": {
      "name": "Juozas Petryla"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/builds": {
      "get": {
        "tags": [
          "Komplektai"
        ],
        "summary": "Gauti komplektų sąrašą",
        "description": "Grąžina komplektus su komponentais, jų pardavėjų pasiūlymais ir atsiliepimais.",
        "operationId": "listBuilds",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "public_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Public Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2147483647,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BuildRead"
                  },
                  "title": "Response Listbuilds"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Komplektai"
        ],
        "summary": "Sukurti kompiuterio komplektą",
        "description": "Sukuria komplektą; jam priklausantys komponentai kuriami per komponentų API.",
        "operationId": "createBuild",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildRead"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "422": {
            "description": "Neteisingas užklausos turinys."
          }
        }
      }
    },
    "/api/v1/builds/{build_id}": {
      "get": {
        "tags": [
          "Komplektai"
        ],
        "summary": "Gauti komplektą",
        "description": "Sudėtinis resursas: viename atsakyme grąžina komplektą, komponentus, jų pardavėjų pasiūlymus ir komplekto atsiliepimus.",
        "operationId": "getBuild",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Build Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildRead"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "404": {
            "description": "Komplektas nerastas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Komplektai"
        ],
        "summary": "Atnaujinti komplektą",
        "description": "Tik savininkas gali pakeisti komplekto metaduomenis. Komponentai ir atsiliepimai tvarkomi atskirais metodais.",
        "operationId": "replaceBuild",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Build Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildReplace"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildRead"
                }
              }
            }
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "404": {
            "description": "Komplektas nerastas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Neteisingas turinys."
          }
        }
      },
      "delete": {
        "tags": [
          "Komplektai"
        ],
        "summary": "Pašalinti komplektą",
        "description": "Savininkas šalina savo komplektą; moderatorius ir admin gali šalinti svetimą viešą komplektą. Kartu pašalinami jo komponentai, pasiūlymai ir atsiliepimai.",
        "operationId": "deleteBuild",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Build Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Reikalingas galiojantis access žetonas."
          },
          "403": {
            "description": "Nepakanka rolės arba įrašo nuosavybės teisių."
          },
          "404": {
            "description": "Komplektas nerastas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BuildCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 3,
            "title": "Name",
            "examples": [
              "1440p Gaming PC"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "description": "Viešą komplektą gali skaityti visi prisijungę naudotojai; privatų – savininkas ir administratorius.",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "BuildCreate"
      },
      "BuildLinks": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self"
          },
          "components": {
            "type": "string",
            "title": "Components"
          },
          "reviews": {
            "type": "string",
            "title": "Reviews"
          }
        },
        "type": "object",
        "required": [
          "self",
          "components",
          "reviews"
        ],
        "title": "BuildLinks"
      },
      "BuildRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/ComponentRead"
            },
            "type": "array",
            "title": "Components"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewRead"
            },
            "type": "array",
            "title": "Reviews"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "links": {
            "$ref": "#/components/schemas/BuildLinks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "owner_id",
          "owner_name",
          "description",
          "is_public",
          "components",
          "reviews",
          "created_at",
          "updated_at",
          "links"
        ],
        "title": "BuildRead"
      },
      "BuildReplace": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 3,
            "title": "Name",
            "examples": [
              "1440p Gaming PC"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "description": "Viešą komplektą gali skaityti visi prisijungę naudotojai; privatų – savininkas ir administratorius.",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "BuildReplace"
      },
      "ComponentCategory": {
        "type": "string",
        "enum": [
          "cpu",
          "motherboard",
          "memory",
          "gpu",
          "storage",
          "psu",
          "case",
          "cooler"
        ],
        "title": "ComponentCategory"
      },
...
      "ComponentLinks": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self"
          },
          "build": {
            "type": "string",
            "title": "Build"
          },
          "offers": {
            "type": "string",
            "title": "Offers"
          }
        },
        "type": "object",
        "required": [
          "self",
          "build",
          "offers"
        ],
        "title": "ComponentLinks"
      },
      "ComponentRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "build_id": {
            "type": "integer",
            "title": "Build Id"
          },
          "category": {
            "$ref": "#/components/schemas/ComponentCategory"
          },
          "manufacturer": {
            "type": "string",
            "title": "Manufacturer"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "specifications": {
            "additionalProperties": true,
            "type": "object",
            "title": "Specifications"
          },
          "offers": {
            "items": {
              "$ref": "#/components/schemas/OfferRead"
            },
            "type": "array",
            "title": "Offers"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "links": {
            "$ref": "#/components/schemas/ComponentLinks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "build_id",
          "category",
          "manufacturer",
          "model",
          "description",
          "specifications",
          "offers",
          "created_at",
          "updated_at",
          "links"
        ],
        "title": "ComponentRead"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OfferLinks": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self"
          },
          "component": {
            "type": "string",
            "title": "Component"
          },
          "build": {
            "type": "string",
            "title": "Build"
          }
        },
        "type": "object",
        "required": [
          "self",
          "component",
          "build"
        ],
        "title": "OfferLinks"
      },
      "OfferRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "component_id": {
            "type": "integer",
            "title": "Component Id"
          },
          "retailer": {
            "type": "string",
            "title": "Retailer"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price"
          },
          "product_url": {
            "type": "string",
            "title": "Product Url"
          },
          "in_stock": {
            "type": "boolean",
            "title": "In Stock"
          },
          "links": {
            "$ref": "#/components/schemas/OfferLinks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "component_id",
          "retailer",
          "price",
          "product_url",
          "in_stock",
          "links"
        ],
        "title": "OfferRead"
      },
...
      "ReviewLinks": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self"
          },
          "build": {
            "type": "string",
            "title": "Build"
          }
        },
        "type": "object",
        "required": [
          "self",
          "build"
        ],
        "title": "ReviewLinks"
      },
      "ReviewRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "build_id": {
            "type": "integer",
            "title": "Build Id"
          },
          "author_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Id"
          },
          "author_name": {
            "type": "string",
            "title": "Author Name"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "links": {
            "$ref": "#/components/schemas/ReviewLinks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "build_id",
          "author_id",
          "author_name",
          "rating",
          "comment",
          "created_at",
          "links"
        ],
        "title": "ReviewRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}